requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv
pydantic
pydantic-settings
jinja2
pyjwt
